requirements.txt
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.3
Django==4.2.1
django-isbn-field==0.5.3
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
Faker==18.10.0
importlib-resources==5.12.0
inflection==0.5.1
isort==5.12.0
jsonschema==4.17.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-stdnum==1.18
pytz==2023.3
PyYAML==6.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.5.0
uritemplate==4.1.1
zipp==3.15.0
